// types/investingInPropeties.types.ts
export type INVESTINGINPROPERTY_RESPONSE = {
  _id: string;
  title: string;
  description: string;
  imageSrc: string;
  isImageRightSide: boolean;
};

export type INVESTINGINPROPERTY_MODEL = {
  _id: string;
  title: string;
  description: string;
  imageSrc: string;
  isImageRightSide: boolean;
};

export type INVESTINGINPROPERTY_FORM_FIELDS = {
  _id: string;
  title: string;
  description: string;
  imageSrc: string;
  imageInput: string;
  isImageRightSide: boolean;
};